+++ src/manager/backend_source.rs
@@ -32,6 +32,6 @@ pub trait BackendSource: Sound {
     fn set_output_channel_count_and_sample_rate(
         &mut self,
         output_channel_count: u16,
-        output_sample_rate: u32,
+        output_sample_rate: u64,
     );
 }
