+++ src/sounds/decoders/mp3.rs
@@ -65,7 +65,7 @@ where
         self.channel_count
     }
 
-    fn sample_rate(&self) -> u32 {
+    fn sample_rate(&self) -> u64 {
         self.sample_rate
     }
