+++ src/sounds/decoders/wav.rs
@@ -50,7 +50,7 @@ where
         self.channel_count
     }
 
-    fn sample_rate(&self) -> u32 {
+    fn sample_rate(&self) -> u64 {
         self.sample_rate
     }
