+++ src/sounds/empty.rs
@@ -1,12 +1,12 @@
 /// A Sound that immediately and always returns Finished
 pub struct Empty {
     channel_count: u16,
-    sample_rate: u32,
+    sample_rate: u64,
 }
 
 impl Empty {
     /// Create a new sound that will immediately and always returns Finished
-    pub fn new(channel_count: u16, sample_rate: u32) -> Empty {
+    pub fn new(channel_count: u16, sample_rate: u64) -> Empty {
         Empty {
             channel_count,
             sample_rate,
@@ -19,7 +19,7 @@ impl crate::Sound for Empty {
         self.channel_count
     }
 
-    fn sample_rate(&self) -> u32 {
+    fn sample_rate(&self) -> u64 {
         self.sample_rate
     }
