+++ src/sounds/wrappers/channel_count_converter.rs
@@ -75,7 +75,7 @@ where
         self.to_count
     }
 
-    fn sample_rate(&self) -> u32 {
+    fn sample_rate(&self) -> u64 {
         self.inner.sample_rate()
     }
