+++ src/tests/sound.rs
@@ -5,7 +5,7 @@ use super::*;
 #[test]
 fn test_skip() {
     {
-        let mut sound = Sawtooth::new(1, std::u16::MAX as u32);
+        let mut sound = Sawtooth::new(1, std::u16::MAX as u64);
         sound.skip(Duration::from_millis(500)).unwrap();
         assert_eq!(
             sound.next_sample().unwrap(),
@@ -13,7 +13,7 @@ fn test_skip() {
         );
     }
     {
-        let mut sound = Sawtooth::new(1, std::u16::MAX as u32);
+        let mut sound = Sawtooth::new(1, std::u16::MAX as u64);
         sound.skip(Duration::from_millis(1000)).unwrap();
         assert_eq!(sound.next_sample().unwrap(), NextSample::Sample(-1));
     }
